GO term correlation analysis

+++ countGO.sh
@@ -10,7 +10,8 @@ download_seqs()
     goterms=$2
     goterms=\"$goterms\"
 
-    wget -o /dev/null -O $species.fa 'http://www.ensembl.org/biomart/martservice?query=<?xml version="1.0" encoding="UTF-8"?> <!DOCTYPE Query> <Query virtualSchemaName = "default" formatter = "FASTA" header = "0" uniqueRows = "0" count = "" datasetConfigVersion = "0.6" completionStamp = "1"> <Dataset name = '$dataset' interface = "default" > <Filter name = "go_parent_term" value = '$goterms'/> <Attribute name = "peptide" /> <Attribute name = "ensembl_peptide_id" /> <Attribute name = "ensembl_gene_id" /> <Attribute name = "ensembl_transcript_id" /> <Attribute name = "gene_biotype" /> <Attribute name = "external_gene_name" /> <Attribute name = "description" /> </Dataset> </Query>'
+    wget -o /dev/null -O $species.fa 'http://www.ensembl.org/biomart/martservice?query=<?xml version="1.0" encoding="UTF-8"?> <!DOCTYPE Query> <Query virtualSchemaName = "default" formatter = "TSV" header = "0" uniqueRows = "1" count = "" datasetConfigVersion = "0.6"> <Dataset name = '$dataset' interface = "default" > <Filter name = "go_parent_term" value = '$goterms'/> <Attribute name = "ensembl_gene_id" /> </Dataset> </Query>'
+    #wget -o /dev/null -O $species.fa 'http://www.ensembl.org/biomart/martservice?query=<?xml version="1.0" encoding="UTF-8"?> <!DOCTYPE Query> <Query virtualSchemaName = "default" formatter = "FASTA" header = "0" uniqueRows = "0" count = "" datasetConfigVersion = "0.6" completionStamp = "1"> <Dataset name = '$dataset' interface = "default" > <Filter name = "go_parent_term" value = '$goterms'/> <Attribute name = "peptide" /> <Attribute name = "ensembl_peptide_id" /> <Attribute name = "ensembl_gene_id" /> <Attribute name = "ensembl_transcript_id" /> <Attribute name = "gene_biotype" /> <Attribute name = "external_gene_name" /> <Attribute name = "description" /> </Dataset> </Query>'
     mv $species.fa count/$species.fa
 }
 
@@ -24,22 +25,29 @@ convertsecs()
 
 start=$SECONDS
 mkdir -p count
+mkdir -p countData
 
 queryid=$2
 
+if [ -e countData/count\_$queryid.txt ]; then 
+    exit 1
+fi
+
 input="data/dataSpecies.txt"
 while read -r species; do
-    echo "Reading data for "$species"."
+    # echo "Reading data for "$species"."
 
     goterms=$1
     dataset=$(python get_dataset.py $species)
     download_seqs $dataset $goterms $species
-    numTerms=$(grep -o ">" count/$species.fa | wc -l)
-    echo "${numTerms}" >> count\_$queryid.txt
+    numTerms=$(wc -l count/$species.fa | awk '{print $1}')
+
+    echo "${numTerms}" >> countData/count\_$queryid.txt
+
 done < "$input"
 
 rm -rf count
 
 end=$SECONDS
 elapsed=$(( end - start ))
-convertsecs $elapsed
+#convertsecs $elapsed
